+++ src/test/integration/utils/commands.js
@@ -21,7 +21,7 @@ function getDocument(dbg) {
   return dbg.win.document;
 }
 
-const { invokeInTab } = require("./mocha");
+const { invokeInTab, selectMenuItem } = require("./mocha");
 
 const { selectors, findSource, getSelector, info } = require("./shared");
 const {
@@ -207,17 +207,6 @@ async function rightClickElement(dbg, elementName, ...args) {
   const el = dbg.win.document.querySelector(selector);
   info('right click on the gutter', el)
   rightClickEl(dbg.win, el);
-}
-
-function selectMenuItem(dbg, index) {
-  const doc =  getDocument(dbg);
-
-  info('select menu item', doc)
-  const popup = doc.querySelector("menupopup[menu-api=\"true\"]");
-  info('found popup',popup)
-  const item = popup.querySelector(`menuitem:nth-child(${index})`);
-  info('found item', item)
-  item.click();
 }
 
 function type(dbg, el, string) {
